Refactor movie feature models and add generic ErrorScreen
- Relocate and consolidate `FilterMenuItem`, `FilterType`, `MovieCardInfo`, and `MoviesState` into a single `Model.kt` file within the `ui.screens.movies` package.
- Update imports across the `:feature-movies` module to reflect the model relocation.
- Create a new `ErrorScreen` composable in the `:designsystem` module to display error messages with an optional retry button.
- Add a "retry" string resource to the design system's `strings.xml`.
- Clean up redundant and unused imports in `MoviesViewModel.kt` and `MoviesIntent.kt`.
