+++ src/feature/feature-movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/movies/components/MoviesHandler.kt
@@ -1,6 +1,6 @@
 import com.gabrielbmoro.moviedb.domain.MoviesRepository
 import com.gabrielbmoro.moviedb.domain.entities.Movie
-import com.gabrielbmoro.moviedb.movies.model.FilterType
+import com.gabrielbmoro.moviedb.movies.ui.screens.movies.FilterType
 import kotlin.collections.List
 
 class MoviesHandler(
