--- src/feature/feature-movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/movies/model/FilterMenuItem.kt
@@ -1,6 +1,0 @@
-package com.gabrielbmoro.moviedb.movies.model
-
-data class FilterMenuItem(
-    val selected: Boolean,
-    val type: FilterType,
-)
