--- src/feature/feature-movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/movies/model/FilterType.kt
@@ -1,8 +1,0 @@
-package com.gabrielbmoro.moviedb.movies.model
-
-enum class FilterType {
-    NowPlaying,
-    TopRated,
-    Popular,
-    UpComing,
-}
