--- src/feature/feature-movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/movies/ui/screens/movies/MovieCardInfo.kt
@@ -1,7 +1,0 @@
-package com.gabrielbmoro.moviedb.movies.ui.screens.movies
-
-data class MovieCardInfo(
-    val movieId: Long,
-    val movieTitle: String,
-    val moviePosterUrl: String,
-)
