+++ src/feature/feature-movies/src/commonMain/kotlin/com/gabrielbmoro/moviedb/movies/ui/screens/movies/MoviesViewModel.kt
@@ -1,12 +1,8 @@
 package com.gabrielbmoro.moviedb.movies.ui.screens.movies
 
 import MoviesHandler
-import com.gabrielbmoro.moviedb.domain.HttpException
 import com.gabrielbmoro.moviedb.domain.entities.Movie
 import com.gabrielbmoro.moviedb.logging.LoggerHelper
-import com.gabrielbmoro.moviedb.movies.model.FilterMenuItem
-import com.gabrielbmoro.moviedb.movies.model.FilterType
-import com.gabrielbmoro.moviedb.movies.model.MoviesState
 import com.gabrielbmoro.moviedb.platform.paging.PagingController
 import com.gabrielbmoro.moviedb.platform.paging.SimplePaging
 import com.gabrielbmoro.moviedb.platform.viewmodel.BaseViewModel
